+++ example/lib/pspdfkit_annotation_processing_example.dart
@@ -14,6 +14,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:pspdfkit_example/platform_utils.dart';
 
 import 'package:pspdfkit_flutter/src/main.dart';
 import 'package:pspdfkit_flutter/src/widgets/pspdfkit_widget_controller.dart';
@@ -64,7 +65,7 @@ class _PspdfkitAnnotationProcessingExampleWidgetState
       'configuration': widget.configuration
     };
 
-    if (defaultTargetPlatform == TargetPlatform.iOS) {
+    if (PlatformUtils.isIOS()) {
       return CupertinoPageScaffold(
           navigationBar: const CupertinoNavigationBar(),
           child: SafeArea(
@@ -117,7 +118,7 @@ class _PspdfkitAnnotationProcessingExampleWidgetState
                       child: Text('Print Annotations'))
                 ]))
               ])));
-    } else if (defaultTargetPlatform == TargetPlatform.android) {
+    } else if (PlatformUtils.isAndroid()) {
       // This example is only supported in iOS at the moment.
       // Support for Android is coming soon.
       return Text('Unsupported Widget');