+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 ## Newest Release
 
+### 3.2.2 - 16 Mar 2022
+
+- Improves the example project by using the `PlatformUtils` class to check for supported platforms (#33212)
+- Adds a new **Save As** example to the example project. (#33376)
+- Updates for PSPDFKit 11.3.0 for iOS. (#33514)
+
+## Previous Releases
+
 ### 3.2.1 - 04 Mar 2022
 
 - Updates for PSPDFKit 8.1.2 for Android. (#33314)
 - Updates for PSPDFKit 11.2.4 for iOS. (#33314)
-
-## Previous Releases
 
 ### 3.2.0 - 14 Feb 2022